PullRequest: mx/970

+++ mx_javamodules.py
@@ -783,7 +783,25 @@ def make_java_module(dist, jdk, javac_daemon=None):
                     jmod_path = jmd.get_jmod_path(respect_stripping=False)
                     if exists(jmod_path):
                         os.remove(jmod_path)
-                    mx.run([jdk.exe_path('jmod'), 'create', '--class-path=' + dest_dir, jmod_path])
+
+                    jdk_jmod = join(jdk_jmods, basename(jmod_path))
+                    jmod_args = ['create', '--class-path=' + dest_dir]
+                    if exists(jdk_jmod):
+                        with ZipFile(jdk_jmod, 'r') as zf:
+                            # Copy commands and legal notices (if any) from JDK version of the module
+                            for jmod_dir, jmod_option in (('bin', '--cmds'), ('legal', '--legal-notices')):
+                                entries = [name for name in zf.namelist() if name.startswith(jmod_dir + '/')]
+                                if entries:
+                                    extracted_dir = join(dest_dir, jmod_dir)
+                                    assert not exists(extracted_dir), extracted_dir
+                                    zf.extractall(dest_dir, entries)
+                                    entries_dir = mx._derived_path(dest_dir, '.' + jmod_dir)
+                                    if exists(entries_dir):
+                                        shutil.rmtree(entries_dir)
+                                    print(extracted_dir, entries_dir)
+                                    os.rename(extracted_dir, entries_dir)
+                                    jmod_args.extend([jmod_option, join(entries_dir)])
+                    mx.run([jdk.exe_path('jmod')] + jmod_args + [jmod_path])
 
                     if tmp_services_dir:
                         os.rename(tmp_services_dir, services_dir)
@@ -800,25 +818,28 @@ def make_java_module(dist, jdk, javac_daemon=None):
                         module_info_class = join(dest_dir, 'module-info.class')
                         zf.write(module_info_class, module_info_arc_dir + basename(module_info_class))
 
-                with mx.Timer('cleanup@' + version, times):
-                    if unversioned_resources_backup:
-                        for dst, contents in unversioned_resources_backup.items():
-                            with open(dst, 'wb') as fp:
-                                fp.write(contents)
-                    for dirpath, dirnames, filenames in os.walk(dest_dir, topdown=False):
-                        del_dirpath = True
-                        for filename in filenames:
-                            abs_filename = join(dirpath, filename)
-                            rel_filename = os.path.relpath(abs_filename, dest_dir)
-                            if rel_filename not in unversioned_resources:
-                                os.remove(abs_filename)
-                            else:
-                                del_dirpath = False
-                        for dname in dirnames:
-                            if exists(join(dirpath, dname)):
-                                del_dirpath = False
-                        if del_dirpath:
-                            os.rmdir(dirpath)
+                if version != max_version:
+                    # Leave output in place for last version so that the jmod command above
+                    # can be manually re-executed (helps debugging).
+                    with mx.Timer('cleanup@' + version, times):
+                        if unversioned_resources_backup:
+                            for dst, contents in unversioned_resources_backup.items():
+                                with open(dst, 'wb') as fp:
+                                    fp.write(contents)
+                        for dirpath, dirnames, filenames in os.walk(dest_dir, topdown=False):
+                            del_dirpath = True
+                            for filename in filenames:
+                                abs_filename = join(dirpath, filename)
+                                rel_filename = os.path.relpath(abs_filename, dest_dir)
+                                if rel_filename not in unversioned_resources:
+                                    os.remove(abs_filename)
+                                else:
+                                    del_dirpath = False
+                            for dname in dirnames:
+                                if exists(join(dirpath, dname)):
+                                    del_dirpath = False
+                            if del_dirpath:
+                                os.rmdir(dirpath)
 
             if files_to_remove:
                 with mx.Timer('cleanup', times), mx.SafeFileCreation(moduleJar) as sfc:
